Adding controllers authz

+++ src/Umbraco.Cms.Api.Management/Controllers/UserGroup/BulkDeleteUserGroupsController.cs
@@ -1,10 +1,12 @@
 ﻿using Asp.Versioning;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Umbraco.Cms.Api.Management.ViewModels.UserGroup;
 using Umbraco.Cms.Core;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
+using Umbraco.Cms.Web.Common.Authorization;
 
 namespace Umbraco.Cms.Api.Management.Controllers.UserGroup;
 
@@ -12,10 +14,12 @@ namespace Umbraco.Cms.Api.Management.Controllers.UserGroup;
 public class BulkDeleteUserGroupsController : UserGroupControllerBase
 {
     private readonly IUserGroupService _userGroupService;
+    private readonly IAuthorizationService _authorizationService;
 
-    public BulkDeleteUserGroupsController(IUserGroupService userGroupService)
+    public BulkDeleteUserGroupsController(IUserGroupService userGroupService, IAuthorizationService authorizationService)
     {
         _userGroupService = userGroupService;
+        _authorizationService = authorizationService;
     }
 
     [HttpDelete]
@@ -24,6 +28,14 @@ public class BulkDeleteUserGroupsController : UserGroupControllerBase
     [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
     public async Task<IActionResult> BulkDelete(DeleteUserGroupsRequestModel model)
     {
+        AuthorizationResult authorizationResult = await _authorizationService.AuthorizeAsync(User, model.UserGroupIds,
+            $"New{AuthorizationPolicies.UserBelongsToUserGroupInRequest}");
+
+        if (!authorizationResult.Succeeded)
+        {
+            return Forbid();
+        }
+
         Attempt<UserGroupOperationStatus> result = await _userGroupService.DeleteAsync(model.UserGroupIds);
 
         return result.Success
