+++ src/Umbraco.Cms.Api.Management/Controllers/UserGroup/DeleteUserGroupController.cs
@@ -1,9 +1,11 @@
 ﻿using Asp.Versioning;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Umbraco.Cms.Core;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
+using Umbraco.Cms.Web.Common.Authorization;
 
 namespace Umbraco.Cms.Api.Management.Controllers.UserGroup;
 
@@ -11,10 +13,12 @@ namespace Umbraco.Cms.Api.Management.Controllers.UserGroup;
 public class DeleteUserGroupController : UserGroupControllerBase
 {
     private readonly IUserGroupService _userGroupService;
+    private readonly IAuthorizationService _authorizationService;
 
-    public DeleteUserGroupController(IUserGroupService userGroupService)
+    public DeleteUserGroupController(IUserGroupService userGroupService, IAuthorizationService authorizationService)
     {
         _userGroupService = userGroupService;
+        _authorizationService = authorizationService;
     }
 
     [HttpDelete("{id:guid}")]
@@ -23,6 +27,14 @@ public class DeleteUserGroupController : UserGroupControllerBase
     [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
     public async Task<IActionResult> Delete(Guid id)
     {
+        AuthorizationResult authorizationResult = await _authorizationService.AuthorizeAsync(User, new[] { id },
+            $"New{AuthorizationPolicies.UserBelongsToUserGroupInRequest}");
+
+        if (!authorizationResult.Succeeded)
+        {
+            return Forbid();
+        }
+
         Attempt<UserGroupOperationStatus> result = await _userGroupService.DeleteAsync(id);
 
         return result.Success
